// src/app/model/option/option-color.model.ts
interface IOptionColor {
    optionColorId: string;
    name: string;
    description: string;
    hexadecimal: string;
    componentName: string;
}

export class OptionColor implements IOptionColor {
    optionColorId: string;
    name: string;
    description: string;
    hexadecimal: string;
    componentName: string;
}